BattleUiEditor: refactored options popup code
moved all code related to options popup into its own script
created a prefab for options popup
