+++ Assets/MenuUi/Prefabs/Settings/BattleUiEditor/OptionsPopup.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e3cd129a57a80a045bedeff408826689
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
