+++ Assets/MenuUi/Scripts/Settings/BattleUiEditor/OptionsPopup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de0d7b787edea4b47ba82caaf176aeba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
